MFS + Modularity
Implemented support for modules and ramdisk. Still porting drivers to
the new system.

MFS:
Implemented WriteFile + SeekFile. I'm waiting with the rest of functions
due to them not being needed atm.

+++ kernel/Include/FileSystems/Mfs.h
@@ -147,7 +147,7 @@ typedef struct _MfsFile
 	/* Information */
 	MString_t *Name;
 	
-	uint32_t Flags;
+	uint16_t Flags;
 	uint32_t DataBucket;
 	uint64_t Size;
 	uint64_t AllocatedSize;
